pyproject.toml
[project]
name = "dxd-rating-system"
version = "0.1.0"
description = "Discord bot for DragxDrive rating management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16.0",
  "discord.py>=2.5.2",
  "psycopg[binary]>=3.2.9",
  "pydantic-settings>=2.10.1",
  "sqlalchemy>=2.0.41",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.15.0",
  "pytest>=8.4.0",
  "ruff>=0.11.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.12"
mypy_path = "src"
packages = ["dxd_rating"]
plugins = ["pydantic.mypy"]
check_untyped_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=80.0.0", "wheel"]
build-backend = "setuptools.build_meta"
